tests/krate/publish: Extract `features` module
